+++ examples/cttc-3gpp-channel-simple-ran.cc
@@ -213,7 +213,7 @@ main (int argc, char *argv[])
     allBwps = CcBwpCreator::GetAllBwps ({band1});
 
     // Beamforming method
-    idealBeamformingHelper->SetAttribute ("IdealBeamformingMethod", TypeIdValue (DirectPathBeamforming::GetTypeId ()));
+    idealBeamformingHelper->SetAttribute ("BeamformingMethod", TypeIdValue (DirectPathBeamforming::GetTypeId ()));
 
     // Antennas for all the UEs
     nrHelper->SetUeAntennaAttribute ("NumRows", UintegerValue (2));