+++ examples/cttc-nr-notching.cc
@@ -336,7 +336,7 @@ main (int argc, char *argv[])
     }
 
   // Beamforming method
-  idealBeamformingHelper->SetAttribute ("IdealBeamformingMethod", TypeIdValue (DirectPathBeamforming::GetTypeId ()));
+  idealBeamformingHelper->SetAttribute ("BeamformingMethod", TypeIdValue (DirectPathBeamforming::GetTypeId ()));
 
   /*
    * Setup the configuration of the spectrum. One operation band is deployed