+++ model/nr-gnb-phy.cc
@@ -789,6 +789,12 @@ NrGnbPhy::StartSlot (const SfnSf &startSlot)
                   NS_LOG_INFO ("Schedule UL CTRL at " << start);
                   Simulator::Schedule (start, &NrGnbPhy::UlCtrl, this, alloc.m_dci);
                 }
+              else if (alloc.m_dci->m_type == DciInfoElementTdma::SRS && alloc.m_dci->m_format == DciInfoElementTdma::UL)
+                {
+                  Time start = GetSymbolPeriod () * alloc.m_dci->m_symStart;
+                  NS_LOG_INFO ("Schedule UL SRS at " << start);
+                  Simulator::Schedule (start, &NrGnbPhy::UlSrs, this, alloc.m_dci);
+                }
             }
         }
     }
@@ -1289,10 +1295,14 @@ NrGnbPhy::UlSrs (const std::shared_ptr<DciInfoElementTdma> &dci)
 
   Time varTtiPeriod = GetSymbolPeriod () * dci->m_numSym;
 
+  m_spectrumPhy->AddExpectedSrsRnti (dci->m_rnti);
+
+/*
   m_spectrumPhy->AddExpectedTb (dci->m_rnti, dci->m_ndi, dci->m_tbSize, dci->m_mcs,
                                 FromRBGBitmaskToRBAssignment (dci->m_rbgBitmask),
                                 dci->m_harqProcess, dci->m_rv, false,
                                 dci->m_symStart, dci->m_numSym, m_currentSlot);
+*/
 
   bool found = false;
   for (uint8_t i = 0; i < m_deviceMap.size (); i++)