+++ model/realistic-beamforming-algorithm.cc
@@ -58,7 +58,7 @@ TypeId
 RealisticBeamformingAlgorithm::GetTypeId (void)
 {
   static TypeId tid = TypeId ("ns3::RealisticBeamformingAlgorithm")
-                     .SetParent<Object> ()
+                     .SetParent<IdealBeamformingAlgorithm> ()
                      .AddConstructor<RealisticBeamformingAlgorithm> ()
                      .AddAttribute ("BeamSearchAngleStep",
                                     "Angle step when searching for the best beam",