+++ test/nr-test-timings.cc
@@ -788,7 +788,7 @@ NrTimingsTest::DoRun (void)
 
   allBwps = CcBwpCreator::GetAllBwps ({band1});
   // Beamforming method
-  idealBeamformingHelper->SetAttribute ("IdealBeamformingMethod", TypeIdValue (DirectPathBeamforming::GetTypeId ()));
+  idealBeamformingHelper->SetAttribute ("BeamformingMethod", TypeIdValue (DirectPathBeamforming::GetTypeId ()));
 
   // Core latency
   epcHelper->SetAttribute ("S1uLinkDelay", TimeValue (MilliSeconds (0)));